requirements.txt
click==7.1.2
Flask==1.1.2
google-api-python-client==1.7.11
google-auth==1.7.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.14.0
idna==2.8
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
matplotlib==3.2.1
more-itertools==7.2.0
packaging==19.2
protobuf==3.12.1
pyparsing==2.4.2
pytest==5.2.2
requests==2.22.0
virtualenv==16.7.7
setuptools~=47.1.1
Pillow==7.1.2
pika~=1.1.0
PyYAML==5.3.1
numpy~=1.18.4